requirements.txt
asgiref==3.8.1
Django==4.2.16
djangorestframework==3.14.0
sqlparse==0.4.3